chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/animation-timeline-none-expected.txt

This is a testharness.js-based test.
FAIL Animation with animation-timeline:none has no effect value assert_equals: expected "0px" but got "100px"
FAIL Animation with unknown timeline name has no effect value assert_equals: expected "0px" but got "100px"
Harness: the test ran to completion.
